Move SQLOperations.Session into a separate Session class
