+++ sql/src/main/java/io/crate/executor/transport/AlterTableOperation.java
@@ -26,6 +26,7 @@ import com.carrotsearch.hppc.cursors.ObjectObjectCursor;
 import com.google.common.annotations.VisibleForTesting;
 import io.crate.Constants;
 import io.crate.action.FutureActionListener;
+import io.crate.action.sql.Session;
 import io.crate.action.sql.ResultReceiver;
 import io.crate.action.sql.SQLOperations;
 import io.crate.analyze.AddColumnAnalyzedStatement;
@@ -138,7 +139,7 @@ public class AlterTableOperation {
 
             SQLOperations.SQLDirectExecutor sqlDirectExecutor = sqlOperations.createSystemExecutor(
                 null,
-                SQLOperations.Session.UNNAMED,
+                Session.UNNAMED,
                 stmt,
                 1);
             try {
